State price returns BigDecimal

+++ catalog/src/main/java/tech/reactiv/ecommerce/catalog/product/ProductPrice.java
@@ -15,7 +15,7 @@ public class ProductPrice {
         this.price = price;
     }
 
-    public int value() {
-        return price.value().intValue();
+    public BigDecimal value() {
+        return price.value();
     }
 }
